requirements.txt
absl-py==2.0.0
adversarial-robustness-toolbox==1.17.0
anyio==4.2.0
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
cleverhans==4.0.0
cloudpickle==3.0.0
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dm-tree==0.1.8
easydict==1.11
exceptiongroup==1.2.0
filelock==3.13.1
fonttools==4.47.0
fsspec==2023.12.2
gast==0.5.4
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
Jinja2==3.1.2
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mnist==0.2.2
mpmath==1.3.0
networkx==3.2.1
nose==1.3.7
numpy==1.26.3
packaging==23.2
pandas==2.2.0
Pillow==9.3.0
pycodestyle==2.11.1
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
scikit-learn==1.1.3
scipy==1.11.4
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sympy==1.12
tensorflow-probability==0.23.0
threadpoolctl==3.2.0
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchvision==0.16.2+cu118
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.1.0
